+++ src/f32/log.rs
@@ -1,6 +1,6 @@
 use crate::double::NormDouble;
 
-// GENERATE: log::consts f32
+// GENERATE: ln::consts f32
 const SQRT_2: f32 = f32::from_bits(0x3FB504F3); // 1.4142135e0
 const LN_2_HI: f32 = f32::from_bits(0x3F317000); // 6.9311523e-1
 const LN_2_LO: f32 = f32::from_bits(0x3805FDF4); // 3.1946183e-5
@@ -9,7 +9,7 @@ const FRAC_2_3_LO: f32 = f32::from_bits(0x332AAAAB); // 3.973643e-8
 const FRAC_4_10_HI: f32 = f32::from_bits(0x3ECCCCCC); // 3.9999998e-1
 const FRAC_4_10_LO: f32 = f32::from_bits(0x32CCCCCD); // 2.3841858e-8
 
-impl crate::generic::Log for f32 {
+impl crate::generic::Ln for f32 {
     #[inline]
     fn sqrt_2() -> Self {
         SQRT_2
@@ -36,8 +36,8 @@ impl crate::generic::Log for f32 {
     }
 
     #[inline]
-    fn log_special_poly(x: Self) -> Self {
-        // GENERATE: log::log_special_poly f32 4
+    fn ln_special_poly(x: Self) -> Self {
+        // GENERATE: ln::ln_special_poly f32 4
         const K2: f32 = f32::from_bits(0x3F2AAAAA); // 6.666666e-1
         const K4: f32 = f32::from_bits(0x3ECCCD3D); // 4.0000334e-1
         const K6: f32 = f32::from_bits(0x3E921C64); // 2.8537285e-1
@@ -48,8 +48,8 @@ impl crate::generic::Log for f32 {
     }
 
     #[inline]
-    fn log_special_poly_ex(x2: Self) -> Self {
-        // GENERATE: log::log_special_poly_ex f32 3
+    fn ln_special_poly_ex(x2: Self) -> Self {
+        // GENERATE: ln::ln_special_poly_ex f32 3
         const K6: f32 = f32::from_bits(0x3E92495E); // 2.85716e-1
         const K8: f32 = f32::from_bits(0x3E634F16); // 2.2198138e-1
         const K10: f32 = f32::from_bits(0x3E454D62); // 1.92678e-1