+++ src/generic/ln.rs
@@ -1,42 +1,42 @@
 use crate::double::{DenormDouble, NormDouble, SemiDouble};
 use crate::traits::{CastInto as _, Float, Int as _};
 
-pub(crate) trait Log: Float {
+pub(crate) trait Ln: Float {
     fn sqrt_2() -> Self;
     fn ln_2_hi() -> Self;
     fn ln_2_lo() -> Self;
     fn frac_2_3_ex() -> NormDouble<Self>;
     fn frac_4_10_ex() -> NormDouble<Self>;
 
-    /// Calculates `(log(1 + x) - log(1 - x) - 2 * x) / x`
+    /// Calculates `(ln(1 + x) - ln(1 - x) - 2 * x) / x`
     ///
     /// `-0.1716 < x < 0.1716`
-    fn log_special_poly(x: Self) -> Self;
+    fn ln_special_poly(x: Self) -> Self;
 
-    /// Calculates `(log(1 + x) - log(1 - x) - 2 * x - (2/3) * x^3 - 0.4 * x^5) / x`
+    /// Calculates `(ln(1 + x) - ln(1 - x) - 2 * x - (2/3) * x^3 - 0.4 * x^5) / x`
     ///
     /// `-0.1716 < x < 0.1716`
-    fn log_special_poly_ex(x2: Self) -> Self;
+    fn ln_special_poly_ex(x2: Self) -> Self;
 }
 
-pub(crate) fn log<F: Log>(x: F) -> F {
+pub(crate) fn ln<F: Ln>(x: F) -> F {
     let (y, edelta) = x.normalize_arg();
     let yexp = y.raw_exp();
     if yexp == F::RawExp::ZERO {
-        // log(±0) = -inf
+        // ln(±0) = -inf
         F::neg_infinity()
     } else if y.sign() {
-        // x < 0, log(x) = NaN
+        // x < 0, ln(x) = NaN
         F::NAN
     } else if yexp == F::MAX_RAW_EXP {
         // propagate infinity or NaN
         y
     } else {
-        log_inner(y, edelta)
+        ln_inner(y, edelta)
     }
 }
 
-pub(crate) fn log_1p<F: Log>(x: F) -> F {
+pub(crate) fn ln_1p<F: Ln>(x: F) -> F {
     let e = x.raw_exp();
     if e == F::RawExp::ZERO {
         // subnormal or zero, log(1 + x) ~= x
@@ -52,45 +52,45 @@ pub(crate) fn log_1p<F: Log>(x: F) -> F {
         // propagate infinity or NaN
         x
     } else {
-        log_1p_inner(x)
+        ln_1p_inner(x)
     }
 }
 
-/// Calculates `log(x * 2^edelta)`
+/// Calculates `ln(x * 2^edelta)`
 ///
 /// `x` must be normal and positive.
-pub(super) fn log_inner<F: Log>(x: F, edelta: F::Exp) -> F {
+pub(super) fn ln_inner<F: Ln>(x: F, edelta: F::Exp) -> F {
     // Algorithm based on one used by the msun math library:
-    //  * log(1 + r) = p * s + 2 * s
+    //  * ln(1 + r) = p * s + 2 * s
     //  * s = r / (2 + r)
-    //  * p = (log(1 + s) - log(1 - s) - 2 * s) / s
+    //  * p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
 
     // Split x * 2^edelta = 2^k * (1 + r)
     //  - k is an integer
     //  - sqrt(2) / 2 <= 1 + r < sqrt(2)
-    let (k, r) = log_split(x, edelta);
+    let (k, r) = ln_split(x, edelta);
 
     // s = r / (2 + r)
-    // So, log(1 + r) = log(1 + s) - log(1 - s)
+    // So, ln(1 + r) = ln(1 + s) - ln(1 - s)
     let s = r / (F::two() + r);
 
-    // p = (log(1 + s) - log(1 - s) - 2 * s) / s
-    let p = F::log_special_poly(s);
+    // p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
+    let p = F::ln_special_poly(s);
 
-    // t1 = k * log(2)
+    // t1 = k * ln(2)
     let kf: F = k.cast_into();
     let t1_hi = kf * F::ln_2_hi();
     let t1_lo = kf * F::ln_2_lo();
 
-    // log(x) = log(1 + r) + k * log(2) = log(1 + r) + t1
-    // where log(1 + r) = p * s + 2 * s
-    //                  = r - s * (r - p)
-    //                  = r - (0.5 * r^2 - s * (0.5 * r^2 + p))
+    // ln(x) = ln(1 + r) + k * ln(2) = ln(1 + r) + t1
+    // where ln(1 + r) = p * s + 2 * s
+    //                 = r - s * (r - p)
+    //                 = r - (0.5 * r^2 - s * (0.5 * r^2 + p))
     let hr2 = F::half() * r * r;
     (((s * (hr2 + p) + t1_lo) - hr2) + r) + t1_hi
 }
 
-fn log_1p_inner<F: Log>(x: F) -> F {
+fn ln_1p_inner<F: Ln>(x: F) -> F {
     // Calculate xp1 + e = 1 + x, where e is an
     // error term to handle rounding in 1 + x.
     let xp1 = (F::one() + x).purify();
@@ -100,15 +100,15 @@ fn log_1p_inner<F: Log>(x: F) -> F {
         (F::one() - xp1) + x
     };
 
-    // Calculate log(1 + x) = log(xp1 + e)
-    log_hi_lo_inner(xp1, e)
+    // Calculate ln(1 + x) = ln(xp1 + e)
+    ln_hi_lo_inner(xp1, e)
 }
 
-/// Returns `(k, r)` as needed by `log_inner`
+/// Returns `(k, r)` as needed by `ln_inner`
 ///
 /// * `x * 2^edelta = 2^k * (1 + r)`
 /// * `sqrt(2) / 2 <= r + 1 <= sqrt(2)`
-pub(super) fn log_split<F: Log>(x: F, edelta: F::Exp) -> (F::Exp, F) {
+pub(super) fn ln_split<F: Ln>(x: F, edelta: F::Exp) -> (F::Exp, F) {
     // Split x * 2^edelta = 2^k * m
     //  - k is an integer
     //  - 1 <= m < 2
@@ -124,60 +124,60 @@ pub(super) fn log_split<F: Log>(x: F, edelta: F::Exp) -> (F::Exp, F) {
     }
 }
 
-// Calculates log(x_hi + x_lo)
-pub(super) fn log_hi_lo_inner<F: Log>(x_hi: F, x_lo: F) -> F {
+// Calculates ln(x_hi + x_lo)
+pub(super) fn ln_hi_lo_inner<F: Ln>(x_hi: F, x_lo: F) -> F {
     // Algorithm based on one used by the msun math library:
-    //  * log(1 + r) = p * s + 2 * s
+    //  * ln(1 + r) = p * s + 2 * s
     //  * s = r / (2 + r)
-    //  * p = (log(1 + s) - log(1 - s) - 2 * s) / s
+    //  * p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
 
     // Split x_hi = 2^k * (1 + r)
     //  - k is an integer
     //  - sqrt(2) / 2 <= 1 + r < sqrt(2)
     //  - e is an error term
-    let (k, r) = log_split(x_hi, F::Exp::ZERO);
+    let (k, r) = ln_split(x_hi, F::Exp::ZERO);
 
     // Calculate a correction term to handle x_lo:
-    // log(x_hi + x_lo) = log(x_hi) + c
-    // c = log(x_hi + x_lo) - log(x_hi) =
-    //   = log((x_hi + x_lo) / x_hi) =
-    //   = log(1 + x_lo / x_hi) ~= x_lo / x_hi
+    // ln(x_hi + x_lo) = ln(x_hi) + c
+    // c = ln(x_hi + x_lo) - ln(x_hi) =
+    //   = ln((x_hi + x_lo) / x_hi) =
+    //   = ln(1 + x_lo / x_hi) ~= x_lo / x_hi
     let c = x_lo / x_hi;
 
     // s = r / (2 + r)
-    // So, log(1 + r) = log(1 + s) - log(1 - s)
+    // So, ln(1 + r) = ln(1 + s) - ln(1 - s)
     let s = r / (F::two() + r);
 
-    // p = (log(1 + s) - log(1 - s) - 2 * s) / s
-    let p = F::log_special_poly(s);
+    // p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
+    let p = F::ln_special_poly(s);
 
-    // t1 = k * log(2) + c
+    // t1 = k * ln(2) + c
     let kf: F = k.cast_into();
     let t1 = DenormDouble::new(F::ln_2_hi(), F::ln_2_lo())
         .pmul1(kf)
         .ladd(c);
 
-    // log(x) = log(x_hi) + c
-    //        = log(1 + r) + k * log(2) + c
-    //        = log(1 + r) + t1
-    // log(1 + r) = p * s + 2 * s
-    //            = r - s * (r - p)
-    //            = r - (0.5 * r^2 - s * (0.5 * r^2 + p))
+    // ln(x) = ln(x_hi) + c
+    //       = ln(1 + r) + k * ln(2) + c
+    //       = ln(1 + r) + t1
+    // ln(1 + r) = p * s + 2 * s
+    //           = r - s * (r - p)
+    //           = r - (0.5 * r^2 - s * (0.5 * r^2 + p))
     let hr2 = F::half() * r * r;
     (((s * (hr2 + p) + t1.lo()) - hr2) + r) + t1.hi()
 }
 
-/// Calculates log(x * 2^edelta)
-pub(super) fn hi_lo_log_inner<F: Log>(x: F, edelta: F::Exp) -> DenormDouble<F> {
+/// Calculates ln(x * 2^edelta)
+pub(super) fn hi_lo_ln_inner<F: Ln>(x: F, edelta: F::Exp) -> DenormDouble<F> {
     // Algorithm based on one used by the msun math library:
-    //  * log(1 + r) = p * s + 2 * s
+    //  * ln(1 + r) = p * s + 2 * s
     //  * s = r / (2 + r)
-    //  * p = (log(1 + s) - log(1 - s) - 2 * s) / s
+    //  * p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
 
     // Split x * 2^edelta = 2^k * (1 + r)
     //  - k is an integer
     //  - sqrt(2) / 2 <= 1 + r < sqrt(2)
-    let (k, r) = log_split(x, edelta);
+    let (k, r) = ln_split(x, edelta);
 
     // rp2 = 2 + r
     let rp2 = SemiDouble::new_qadd11(F::two(), r);
@@ -186,39 +186,39 @@ pub(super) fn hi_lo_log_inner<F: Log>(x: F, edelta: F::Exp) -> DenormDouble<F> {
     let s = (SemiDouble::new(r) / rp2).to_semi();
     let s2 = s.square().to_semi();
 
-    // p = (log(1 + s) - log(1 - s) - 2 * s) / s
-    let p = hi_lo_log_special_poly(s2).to_semi();
+    // p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
+    let p = hi_lo_ln_special_poly(s2).to_semi();
 
-    // t1 = k * log(2)
+    // t1 = k * ln(2)
     let kf: F = k.cast_into();
     let t1 = DenormDouble::new(F::ln_2_hi(), F::ln_2_lo()).pmul1(kf);
 
-    // t2 = log(1 + r) = p * s + 2 * s
+    // t2 = ln(1 + r) = p * s + 2 * s
     let ps = p * s;
     let twos = s.pmul1(F::two());
     let t2 = twos.to_denorm().qadd2(ps);
 
-    // log(2^k * (1 + r)) = t1 + t2
+    // ln(2^k * (1 + r)) = t1 + t2
     t1.qadd2(t2)
 }
 
-/// Calculates log((x_hi + x_lo) * 2^edelta)
-pub(super) fn hi_lo_log_hi_lo_inner<F: Log>(x: NormDouble<F>, edelta: F::Exp) -> DenormDouble<F> {
+/// Calculates ln((x_hi + x_lo) * 2^edelta)
+pub(super) fn hi_lo_ln_hi_lo_inner<F: Ln>(x: NormDouble<F>, edelta: F::Exp) -> DenormDouble<F> {
     // Algorithm based on one used by the msun math library:
-    //  * log(1 + r) = p * s + 2 * s
+    //  * ln(1 + r) = p * s + 2 * s
     //  * s = r / (2 + r)
-    //  * p = (log(1 + s) - log(1 - s) - 2 * s) / s
+    //  * p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
 
     // Split x_hi * 2^edelta = 2^k * (1 + r)
     //  - k is an integer
     //  - sqrt(2) / 2 <= 1 + r < sqrt(2)
-    let (k, r) = log_split(x.hi(), edelta);
+    let (k, r) = ln_split(x.hi(), edelta);
 
     // Calculate a correction term to handle x_lo:
-    // log(x_hi + x_lo) = log(x_hi) + c
-    // c = log(x_hi + x_lo) - log(x_hi) =
-    //   = log((x_hi + x_lo) / x_hi) =
-    //   = log(1 + x_lo / x_hi) ~= x_lo / x_hi
+    // ln(x_hi + x_lo) = ln(x_hi) + c
+    // c = ln(x_hi + x_lo) - ln(x_hi) =
+    //   = ln((x_hi + x_lo) / x_hi) =
+    //   = ln(1 + x_lo / x_hi) ~= x_lo / x_hi
     let c = x.lo() / x.hi();
 
     // rp2 = 2 + r
@@ -228,30 +228,30 @@ pub(super) fn hi_lo_log_hi_lo_inner<F: Log>(x: NormDouble<F>, edelta: F::Exp) ->
     let s = (SemiDouble::new(r) / rp2).to_semi();
     let s2 = s.square().to_semi();
 
-    // p = (log(1 + s) - log(1 - s) - 2 * s) / s
-    let p = hi_lo_log_special_poly(s2).to_semi();
+    // p = (ln(1 + s) - ln(1 - s) - 2 * s) / s
+    let p = hi_lo_ln_special_poly(s2).to_semi();
 
-    // t1 = k * log(2) + c
+    // t1 = k * ln(2) + c
     let kf: F = k.cast_into();
     let t1 = DenormDouble::new(F::ln_2_hi(), F::ln_2_lo())
         .pmul1(kf)
         .ladd(c);
 
-    // t2 = log(1 + r) = p * s + 2 * s
+    // t2 = ln(1 + r) = p * s + 2 * s
     let ps = p * s;
     let twos = s.pmul1(F::two());
     let t2 = twos.to_denorm().qadd2(ps);
 
-    // log(2^k * (1 + r)) + c = t1 + t2
+    // ln(2^k * (1 + r)) + c = t1 + t2
     t1.qadd2(t2)
 }
 
-/// Calculates `(log(1 + x) - log(1 - x) - 2 * x) / x`
+/// Calculates `(ln(1 + x) - ln(1 - x) - 2 * x) / x`
 ///
 /// `-0.1716 < x < 0.1716`
-fn hi_lo_log_special_poly<F: Log>(x2: SemiDouble<F>) -> DenormDouble<F> {
+fn hi_lo_ln_special_poly<F: Ln>(x2: SemiDouble<F>) -> DenormDouble<F> {
     // p0 = (p - 2/3 * x^2 - 0.4 * x^4) / x^4
-    let p0 = F::log_special_poly_ex(x2.to_single());
+    let p0 = F::ln_special_poly_ex(x2.to_single());
 
     // p1 = (p - 2/3 * x^2) / x^4 = p0 + 0.4
     let p1 = F::frac_4_10_ex().to_denorm().qadd1(p0).to_semi();
@@ -271,38 +271,38 @@ mod tests {
     use crate::FloatMath;
     use crate::traits::Float;
 
-    fn test_log<F: Float + FloatMath>() {
-        use crate::log;
+    fn test_ln<F: Float + FloatMath>() {
+        use crate::ln;
 
-        assert_is_nan!(log(F::NAN));
-        assert_is_nan!(log(-F::one()));
-        assert_is_nan!(log(F::neg_infinity()));
-        assert_total_eq!(log(F::ZERO), F::neg_infinity());
-        assert_total_eq!(log(-F::ZERO), F::neg_infinity());
-        assert_total_eq!(log(F::INFINITY), F::INFINITY);
+        assert_is_nan!(ln(F::NAN));
+        assert_is_nan!(ln(-F::one()));
+        assert_is_nan!(ln(F::neg_infinity()));
+        assert_total_eq!(ln(F::ZERO), F::neg_infinity());
+        assert_total_eq!(ln(-F::ZERO), F::neg_infinity());
+        assert_total_eq!(ln(F::INFINITY), F::INFINITY);
     }
 
-    fn test_log_1p<F: Float + FloatMath>() {
-        use crate::log_1p;
+    fn test_ln_1p<F: Float + FloatMath>() {
+        use crate::ln_1p;
 
-        assert_is_nan!(log_1p(F::NAN));
-        assert_is_nan!(log_1p(-(F::one() + F::half())));
-        assert_is_nan!(log_1p(F::neg_infinity()));
-        assert_total_eq!(log_1p(-F::one()), F::neg_infinity());
-        assert_total_eq!(log_1p(-F::ZERO), -F::ZERO);
-        assert_total_eq!(log_1p(F::ZERO), F::ZERO);
-        assert_total_eq!(log_1p(F::INFINITY), F::INFINITY);
+        assert_is_nan!(ln_1p(F::NAN));
+        assert_is_nan!(ln_1p(-(F::one() + F::half())));
+        assert_is_nan!(ln_1p(F::neg_infinity()));
+        assert_total_eq!(ln_1p(-F::one()), F::neg_infinity());
+        assert_total_eq!(ln_1p(-F::ZERO), -F::ZERO);
+        assert_total_eq!(ln_1p(F::ZERO), F::ZERO);
+        assert_total_eq!(ln_1p(F::INFINITY), F::INFINITY);
     }
 
     #[test]
     fn test_f32() {
-        test_log::<f32>();
-        test_log_1p::<f32>();
+        test_ln::<f32>();
+        test_ln_1p::<f32>();
     }
 
     #[test]
     fn test_f64() {
-        test_log::<f64>();
-        test_log_1p::<f64>();
+        test_ln::<f64>();
+        test_ln_1p::<f64>();
     }
 }